Implementation auth: Login and logout

+++ index.php
@@ -1,6 +1,18 @@
 <?php
 include "bootstrap/init.php";
 
+if(isset($_GET['logout'])){
+     logout();
+}
+
+if(!isLoggedIn()){
+     //Redirect to aut form
+     redirect(site_url('auth.php'));
+}
+
+#user is LoggedIn
+$user = getLoggedInUser();
+
 if(isset($_GET['delete_folder']) && is_numeric($_GET['delete_folder'])){
      deleteFolder($_GET['delete_folder']);
      removeTasks($_GET['delete_folder']);
